Fix some ZFS Test Suite issues
* Add 'zfs bookmark' coverage (zfs_bookmark_cliargs)

 * Add OpenZFS 8166 coverage (zpool_scrub_offline_device)

 * Fix "busy" zfs_mount_remount failures

 * Fix bootfs_003_pos, bootfs_004_neg, zdb_005_pos local cleanup

 * Update usage of $KEEP variable, add get_all_pools() function

 * Enable history_008_pos and rsend_019_pos (non-32bit builders)

 * Enable zfs_copies_005_neg, update local cleanup

 * Fix zfs_send_007_pos (large_dnode + OpenZFS 8199)

 * Fix rollback_003_pos (use dataset name, not mountpoint, to unmount)

 * Update default_raidz_setup() to work properly with more than 3 disks

 * Use $TEST_BASE_DIR instead of hardcoded (/var)/tmp for file VDEVs

 * Update usage of /dev/random to /dev/urandom

Issue #6086 
Closes #5658 
Closes #6143 
Closes #6421 
Closes #6627 
Closes #6632

+++ tests/zfs-tests/tests/functional/pool_names/pool_names_002_neg.ksh
@@ -63,12 +63,10 @@ function cleanup
 
 log_onexit cleanup
 
-typeset exclude=`eval echo \"'(${KEEP})'\"`
-for pool in $(zpool list -H -o name | \
-	egrep -v "$exclude" | \
-	grep -v "$TESTPOOL" | \
-	egrep -v "$NO_POOLS"); do
-	log_must zpool destroy $pool
+for pool in $(get_all_pools); do
+	if [[ "$pool" != "$TESTPOOL" ]]; then
+		log_must zpool destroy $pool
+	fi
 done
 
 DISK=${DISKS%% *}
